Moved test sources under 'src' and 'include' dirs
