+++ test/include/proto_nmea_test/fakes/FakeUtils.h
