+++ test/include/proto_nmea_test/fakes/etc/checksum/ChecksumValueComparatorFakes.h
