+++ test/include/proto_nmea_test/fakes/format/ChecksumValidatorFakes.h
