+++ test/src/etc/checksum/ChecksumCalculatorTests.cpp
@@ -3,10 +3,10 @@
 // Copyright (c) 2019 Takpit. All rights reserved.
 //
 
-#include "../../CatchSmartInclude.hpp"
+#include "CatchSmartInclude.hpp"
 #include <fff.h>
-#include "../../utility/fakes/FakeUtils.h"
-#include "fakes/ChecksumXORCalculatorFakes.h"
+#include "proto_nmea_test/fakes/FakeUtils.h"
+#include "proto_nmea_test/fakes/etc/checksum/ChecksumXORCalculatorFakes.h"
 
 #include <string>
