+++ test/src/etc/checksum/fakes/ChecksumValueComparatorFakes.c
@@ -3,6 +3,6 @@
 // Copyright (c) 2019 Takpit. All rights reserved.
 //
 
-#include "ChecksumValueComparatorFakes.h"
+#include "proto_nmea_test/fakes/etc/checksum/ChecksumValueComparatorFakes.h"
 
 DEFINE_FAKE_VALUE_FUNC(int, compareChecksumData, const unsigned char*, unsigned int)
