+++ test/src/format/fakes/ChecksumFormatValidatorFakes.c
@@ -3,6 +3,6 @@
 // Copyright (c) 2019 Takpit. All rights reserved.
 //
 
-#include "ChecksumFormatValidatorFakes.h"
+#include "proto_nmea_test/fakes/format/ChecksumFormatValidatorFakes.h"
 
 DEFINE_FAKE_VALUE_FUNC(int8_t, validateChecksumFormat, const unsigned char*)
