+++ test/src/format/fakes/ChecksumValidatorFakes.c
@@ -2,6 +2,6 @@
 // Created by timor on 04/08/2019.
 //
 
-#include "ChecksumValidatorFakes.h"
+#include "proto_nmea_test/fakes/format/ChecksumValidatorFakes.h"
 
 DEFINE_FAKE_VALUE_FUNC(int8_t, validateMessageChecksum, const unsigned char*, size_t)
